Ease of use nlrq wrapper function for fitting growth models specified by growthSS

Helper function generally called from fitGrowth.

Usage

fitGrowthnlrq(ss, cores = getOption("mc.cores", 1), ...)

Arguments

ss

A list generated by growthSS.

cores

Optionally specify how many cores to run in parallel if ss$taus is >1L. Defaults to 1 if mc.cores option is not set globally.

...

Additional arguments passed to quantreg::nlrq.

Value

An nlrqModel object if tau is length of 1 or a list of such models for multiple taus, see ?quantreg::nlrq or ?nls::nlsModel for details.