Read in plantCV json keeping either single value, multi value, or all traits.

Read in plantCV json keeping either single value, multi value, or all traits.

Usage

read.pcv.jsn(
  file,
  dtSuffix = "datatype",
  valSuffix = "value",
  labSuffix = "label",
  output = "sv"
)

Arguments

file

Path to the plantCV output json.

dtSuffix

Suffix for columns denoting data type. This will be one of the innermost keys in the json. Default is "datatype"

valSuffix

Suffix for columns denoting phenotype values. This will be one of the innermost keys in the json. Default is "value"

labSuffix

Suffix for columns denoting phenotype labels. This will be one of the innermost keys in the json. Default is "label"

output

What phenotypes should be returned? This can be a vector of names present in "label" or one of the built in options. Built in options are "sv" for single value traits only, "mv" for multi value traits only, and "all" for all traits. Note that by default tuple value data is output with single value data in a wide format.

Value

Returns a dataframe.

Examples

## Not run:

if (FALSE) {
  file <- "path/to/your/json/file.json"
  x <- read.pcv.jsn(file, output = "sv")
  y <- read.pcv.jsn(file, output = c("area", "perimeter"))
  z <- read.pcv.jsn(file, output = "mv")
}

## End(Not run)