Ease of use wrapper function for fitting various growth models specified by growthSS

Ease of use wrapper function for fitting various growth models specified by growthSS

Usage

fitGrowth(ss, ...)

Arguments

ss

A list generated by growthSS.

...

Additional arguments passed to model fitting functions determined by ss$type. For type = "nlme" these are passed to nlme::nlmeControl, not nlme::nlme. Additional arguments are documented in fitGrowthbrms, fitGrowthnlme, fitGrowthnls, fitGrowthnlrq, fitGrowthmgcvgam, fitGrowthsurvreg, fitGrowthflexsurv.

Value

A fit model from the selected type.

Examples

## Not run:

simdf <- growthSim("logistic",
  n = 20, t = 25,
  params = list("A" = c(200, 160), "B" = c(13, 11), "C" = c(3, 3.5))
)
ss <- growthSS(
  model = "logistic", form = y ~ time | group,
  df = simdf, type = "nls"
)
fitGrowth(ss)
#> Nonlinear regression model
#>   model: y ~ A[group]/(1 + exp((B[group] - time)/C[group]))
#>    data: ss[["df"]]
#>      A1      A2      B1      B2      C1      C2 
#> 200.921 162.651  13.040  10.486   3.059   3.476 
#>  residual sum-of-squares: 197187
#> 
#> Number of iterations to convergence: 3 
#> Achieved convergence tolerance: 3.734e-06
ss <- growthSS(
  model = "gam", form = y ~ time | group,
  df = simdf, type = "nls"
)
fitGrowth(ss)
#> 
#> Call:
#> stats::lm(formula = ss[["formula"]], data = ss[["df"]])
#> 
#> Coefficients:
#>      (Intercept)         bs(time)1         bs(time)2         bs(time)3  
#>            8.732           -43.691           225.766           183.537  
#>           groupb  bs(time)1:groupb  bs(time)2:groupb  bs(time)3:groupb  
#>           -1.302            69.074           -38.049           -36.680  
#> 

## End(Not run)